Use u32 rather than u16 for user group limit (#693)

+++ v2/backend/canisters/user/api/src/updates/create_group.rs
@@ -17,7 +17,7 @@ pub enum Response {
     NameTooLong(FieldTooLongResult),
     DescriptionTooLong(FieldTooLongResult),
     AvatarTooBig(FieldTooLongResult),
-    MaxGroupsCreated(u16),
+    MaxGroupsCreated(u32),
     NameTaken,
     Throttled,
     InternalError,
